python/mnemos-client/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mnemos-client"
version = "0.1.0"
description = "Python gRPC client for Mnemos agent memory engine"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "Mnemos" }
]
dependencies = [
  "grpcio>=1.62.0",
  "protobuf>=5.26.0",
]

[project.optional-dependencies]
dev = [
  "grpcio-tools>=1.62.0",
  "mypy>=1.10.0",
]
embeddings = [
  "google-genai>=0.7.0",
  "openai>=1.52.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["mnemos_client*"]
